pyproject.toml
[project]
name = "langchain-basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "streamlit>=1.52.1",
    "watchdog>=6.0.0",
]
